requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1-modules==0.4.2
PyJWT==2.9.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
typing-extensions==4.13.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
geopy>=2.4.1
httpx[http2]>=0.27.2
imagehash>=4.3.1
pillow>=10.4.0
python-dateutil>=2.9.0
rich>=13.9.1
beautifultable>=1.1.0
beautifulsoup4>=4.12.3
alive-progress>=3.1.5
protobuf>=5.28.2
autoslot>=2022.12.1
humanize>=4.10.0
inflection>=0.5.1
jsonpickle>=3.3.0
packaging>=24.1
rich-argparse>=1.5.2
dnspython>=2.7.0
drf-yasg==1.21.7